Moved sceDisplay to a proper module and SyscallHandler is now clean and only doing the dispatching. Code cleanup in many files : reorganized imports, moved some else blocks down one level if the previous if block ended with a return, removed some redundant "this." qualifier, added tons of @Override.
Let's now try to make it almost warning free, so we can find the important warnings easier !

+++ trunk/src/jpcsp/Allegrex/LsuState.java
@@ -50,7 +50,7 @@ public class LsuState extends MduState {
     }
 
     public void doLB(int rt, int rs, int simm16) {
-        int word = (int)(byte)memory.read8(gpr[rs] + simm16);
+        int word = (byte)memory.read8(gpr[rs] + simm16);
         if (rt != 0) {
             gpr[rt] = word;
         }
@@ -72,7 +72,7 @@ public class LsuState extends MduState {
             }
         }
 
-        int word = (int)(short)memory.read16(gpr[rs] + simm16);
+        int word = (short)memory.read16(gpr[rs] + simm16);
         if (rt != 0) {
             gpr[rt] = word;
         }
